// HelpDeskFE/HelpDeskFETp/src/api/model/getApiTicketsParams.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export type GetApiTicketsParams = {
IdUser?: number;
IdTypeError?: number;
IdArea?: number;
IdSoftwareSystem?: number;
IdImpact?: number;
IdPriority?: number;
DateFrom?: string;
DateTo?: string;
Keyword?: string;
PageNumber?: number;
PageSize?: number;
};