// HelpDeskFE/HelpDeskFETp/src/api/model/updateTicketDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface UpdateTicketDto {
  /**
     * @minLength 0
     * @maxLength 360
     */
  description: string;
  /** @minimum 1 */
  idTypeError: number;
  /** @minimum 1 */
  idSoftwareSystem: number;
  /** @minimum 1 */
  idImpact: number;
  /** @minimum 1 */
  idPriority: number;
  /** @minimum 1 */
  idArea: number;
  isActive?: boolean;
}